docker/paddleocr-service/requirements.txt
Flask==3.1.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
PyMuPDF==1.25.5
paddlepaddle==2.6.2
paddleocr==2.8.1
